refactor(coeus): Split provider backend modules
Keep vendor backend manifests small by moving provider identity, reductions, elementwise dispatch, and runtime integration into canonical leaves. Preserve public re-exports and monomorphized kernel boundaries.
